requirements.txt
langchain
langchain-community
langchain-google-genai
chromadb
streamlit
pypdf
python-dotenv
tiktoken
langchain-chroma
